Added ambient light and vulkan recreate swap chain.

+++ src/Platform/Vulkan/Internal/VulkanSwapchain.cpp
@@ -252,6 +252,7 @@ namespace Framework
 
 			return VK_PRESENT_MODE_FIFO_KHR;
 		}
+
 		VkExtent2D VulkanSwapchain::ChooseSwapExtent(const VkSurfaceCapabilitiesKHR& capabilities)
 		{
 			if (capabilities.currentExtent.width != UINT32_MAX)
@@ -262,7 +263,7 @@ namespace Framework
 			{
 				int width, height;
 				glfwGetWindowSize(static_cast<GLFWwindow*>(m_vulkanContext->GetWindow()), &width, &height);
-				VkExtent2D actualExtent = { width, height };
+				VkExtent2D actualExtent = { static_cast<uint32_t>(width), static_cast<uint32_t>(height) };
 
 				actualExtent.width = std::clamp(actualExtent.width, capabilities.minImageExtent.width, capabilities.maxImageExtent.width);
 				actualExtent.height = std::clamp(actualExtent.height, capabilities.minImageExtent.height, capabilities.maxImageExtent.height);
